Move ModalHeader to separate component
