+++ src/signals/incident-management/containers/IncidentDetail/components/StatusForm/components/DefaultTexts/styled.ts
@@ -1,27 +1,5 @@
 import styled, { css } from 'styled-components'
-import {
-  Link,
-  Heading,
-  themeColor,
-  themeSpacing,
-  Button,
-} from '@amsterdam/asc-ui'
-
-export const CloseButton = styled(Button)`
-  border: none;
-  padding: 0;
-  width: 20px;
-  height: 20px;
-`
-
-export const ModalHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: ${themeSpacing(3, 4, 3, 4)};
-  margin: 0;
-  border-bottom: 1px solid ${themeColor('tint', 'level5')};
-`
+import { Link, themeColor, themeSpacing } from '@amsterdam/asc-ui'
 
 export const StyledDefaultText = styled.div<{ empty?: boolean }>`
   margin-bottom: ${themeSpacing(5)};
@@ -31,10 +9,6 @@ export const StyledDefaultText = styled.div<{ empty?: boolean }>`
     css`
       color: ${themeColor('tint', 'level5')};
     `}
-`
-
-export const StyledHeading = styled(Heading)`
-  margin: 0;
 `
 
 export const StyledLink = styled(Link)`
